+++ scrapy_playwright/__init__.py
@@ -1,6 +1,1 @@
 __version__ = "0.0.40"
-
-
-from ._utils import ensure_future  # noqa: F401
-from .handler import ScrapyPlaywrightDownloadHandler  # noqa: F401
-from .page import PageMethod  # noqa: F401
