+++ scrapy_playwright/_utils.py
@@ -162,24 +162,3 @@ class _ThreadedLoopAdapter:
             asyncio.run_coroutine_threadsafe(cls._coro_queue.join(), cls._loop)
             cls._loop.call_soon_threadsafe(cls._loop.stop)
             cls._thread.join()
-
-
-def ensure_future(coro: Awaitable) -> asyncio.Future:
-    """Wrap a coroutine in a Future assigned to the threaded event loop.
-
-    On windows, Playwright runs in an event loop of its own in a separate thread.
-    If Playwright coroutines are awaited directly, they are assigned to the main
-    thread's event loop, resulting in: "ValueError: The future belongs to a
-    different loop than the one specified as the loop argument"
-
-    Usage:
-    ```
-    from playwright.async_api import Page
-    from scrapy_playwright import ensure_future
-
-    async def parse(self, response):
-        page: Page = response.meta["playwright_page"]
-        await ensure_future(page.screenshot(path="example.png", full_page=True))
-    ```
-    """
-    return _ThreadedLoopAdapter._ensure_future(coro)
